Finish implementation of coupling algorithm stores via server configuration
