+++ vantage6-server/vantage6/server/resource/algorithm_store.py
@@ -1,9 +1,8 @@
 import logging
 
-import requests
-from flask import Response, request, g
-from flask_restful import Api
 from http import HTTPStatus
+from flask import request, g
+from flask_restful import Api
 from sqlalchemy import or_
 
 from vantage6.server import db
@@ -12,6 +11,11 @@ from vantage6.server.resource.common.input_schema import AlgorithmStoreInputSche
 from vantage6.server.permission import RuleCollection, Operation as P
 from vantage6.server.resource.common.output_schema import AlgorithmStoreSchema
 from vantage6.server.resource import with_user, ServicesResources
+from vantage6.server.algo_store_communication import (
+    post_algorithm_store,
+    get_server_url,
+    request_algo_store,
+)
 
 
 module_name = __name__.split(".")[-1]
@@ -63,191 +67,6 @@ class AlgorithmStoreBase(ServicesResources):
     def __init__(self, socketio, mail, api, permissions, config):
         super().__init__(socketio, mail, api, permissions, config)
         self.r_col: RuleCollection = getattr(self.permissions, "collaboration")
-
-    @classmethod
-    def _request_algo_store(
-        cls,
-        algo_store_url: str,
-        server_url: str,
-        endpoint: str,
-        method: str,
-        force: bool = False,
-        headers: dict = None,
-    ) -> tuple[dict | Response, HTTPStatus]:
-        """
-        Whitelist this vantage6 server url for the algorithm store.
-
-        Parameters
-        ----------
-        algo_store_url : str
-            URL to the algorithm store
-        server_url : str
-            URL to this vantage6 server. This is used to whitelist this server
-            at the algorithm store.
-        endpoint : str
-            Endpoint to use at the algorithm store.
-        method : str
-            HTTP method to use.
-        force : bool
-            If True, the algorithm store will be added even if the algorithm
-            store url is insecure (i.e. localhost)
-        headers : dict
-            Headers to be included in the request. Usually, these will be Authorization
-            headers
-
-        Returns
-        -------
-        tuple[dict | Response, HTTPStatus]
-            The response of the algorithm store and the HTTP status. If the
-            algorithm store is not reachable, a dict with an error message is
-            returned instead of the response.
-        """
-        # TODO this is not pretty, but it works for now. This should change
-        # when we have a separate auth service
-        is_localhost_algo_store = cls._contains_localhost(algo_store_url)
-        try:
-            response = cls._execute_algo_store_request(
-                algo_store_url, server_url, endpoint, method, force, headers
-            )
-        except requests.exceptions.ConnectionError as exc:
-            if not is_localhost_algo_store:
-                log.warning("Request to algorithm store failed")
-                log.exception(exc)
-            response = None
-
-        if not response and is_localhost_algo_store:
-            # try again with the docker host ip
-            algo_store_url = algo_store_url.replace(
-                "localhost", "host.docker.internal"
-            ).replace("127.0.0.1", "host.docker.internal")
-            try:
-                response = cls._execute_algo_store_request(
-                    algo_store_url, server_url, endpoint, method, force, headers
-                )
-            except requests.exceptions.ConnectionError as exc:
-                log.warning("Request to algorithm store failed")
-                log.exception(exc)
-                response = None
-
-        if response is None:
-            return {
-                "msg": "Algorithm store cannot be reached. Make sure that "
-                "it is online and that you have not included /api at the "
-                "end of the algorithm store URL"
-            }, HTTPStatus.BAD_REQUEST
-        elif response.status_code not in [HTTPStatus.CREATED, HTTPStatus.OK]:
-            try:
-                msg = (
-                    f"Algorithm store error: {response.json()['msg']}, HTTP status: "
-                    f"{response.status_code}"
-                )
-            except KeyError:
-                msg = (
-                    "Communication to algorithm store failed. HTTP status: "
-                    f"{response.status_code}"
-                )
-            return {"msg": msg}, HTTPStatus.BAD_REQUEST
-        # else: server has been registered at algorithm store, proceed
-        return response, response.status_code
-
-    @staticmethod
-    def _contains_localhost(url: str) -> bool:
-        """Check if the url refers to localhost address"""
-        return url.startswith("http://localhost") or url.startswith("http://127.0.0.1")
-
-    # TODO this function and above should be moved to some kind of client lib
-    @staticmethod
-    def _execute_algo_store_request(
-        algo_store_url: str,
-        server_url: str,
-        endpoint: str,
-        method: str,
-        force: bool,
-        headers: dict = None,
-    ) -> requests.Response:
-        """
-        Send a request to the algorithm store to whitelist this vantage6 server
-        url for the algorithm store.
-
-        Parameters
-        ----------
-        algo_store_url : str
-            URL to the algorithm store
-        server_url : str
-            URL to this vantage6 server. This is used to whitelist this server
-            at the algorithm store.
-        endpoint : str
-            Endpoint to use at the algorithm store.
-        method : str
-            HTTP method to use. Choose "post" for adding the server url and
-            "delete" for removing it.
-        force : bool
-            If True, the algorithm store will be added even if the algorithm
-            store url is insecure (i.e. localhost)
-        headers : dict
-            Headers to be included in the request. Usually, these will be Authorization
-            headers
-
-        Returns
-        -------
-        requests.Response | None
-            Response from the algorithm store. If the algorithm store is not
-            reachable, None is returned
-        """
-        if server_url.endswith("/"):
-            server_url = server_url[:-1]
-        if algo_store_url.endswith("/"):
-            algo_store_url = algo_store_url[:-1]
-
-        param_dict = {"url": server_url}
-        if force:
-            param_dict["force"] = True
-
-        # set headers
-        if not headers:
-            headers = {}
-        headers = {"server_url": server_url}
-
-        params = None
-        json = None
-        if method == "get":
-            request_function = requests.get
-            params = param_dict
-        elif method == "post":
-            request_function = requests.post
-            json = param_dict
-        elif method == "delete":
-            request_function = requests.delete
-            params = param_dict
-        else:
-            raise ValueError(f"Method {method} not supported")
-
-        return request_function(
-            f"{algo_store_url}/api/{endpoint}",
-            params=params,
-            json=json,
-            headers=headers,
-        )
-
-    @staticmethod
-    def _get_server_url(config: dict, server_url_from_request: str | None) -> str:
-        """ "
-        Get the server url from the server configuration, or from the request
-        data if it is not present in the configuration.
-
-        Parameters
-        ----------
-        config : dict
-            Server configuration
-        server_url_from_request : str | None
-            Server url from the request data.
-
-        Returns
-        -------
-        str | None
-            The server url
-        """
-        return config.get("server_url", server_url_from_request)
 
     @staticmethod
     def get_authorization_headers_from_request() -> dict:
@@ -486,104 +305,14 @@ class AlgorithmStores(AlgorithmStoreBase):
                 "msg": "You lack the permission to do that!"
             }, HTTPStatus.UNAUTHORIZED
 
-        return self.post_algorithm_store(
+        response, status = post_algorithm_store(
             request.get_json(),
             self.config,
-            self.get_authorization_headers_from_request(),
-        )
-
-    @classmethod
-    def post_algorithm_store(
-        cls, data: dict, config: dict, headers: dict
-    ) -> tuple[dict, HTTPStatus]:
-        """Add algorithm store to a collaboration
-
-        Parameters
-        ----------
-        data : dict
-            Request body as required for POST /algorithmstore request
-        config : dict
-            Server configuration
-        headers : dict
-            Headers to be included in the request. Usually, these will be Authorization
-            headers
-
-        Returns
-        -------
-        tuple[dict, HTTPStatus]
-            Response and HTTP status code
-        """
-        # check if algorithm store is already available for the collaboration
-        collaboration_id = data.get("collaboration_id", None)
-        algorithm_store_url = data["algorithm_store_url"]
-        if algorithm_store_url.endswith("/"):
-            algorithm_store_url = algorithm_store_url[:-1]
-        existing_algorithm_stores = db.AlgorithmStore.get_by_url(algorithm_store_url)
-        records_to_delete = []
-        if existing_algorithm_stores:
-            collabs_with_algo_store = [
-                a.collaboration_id for a in existing_algorithm_stores
-            ]
-            if None in collabs_with_algo_store:
-                return {
-                    "msg": "Algorithm store is already available for all "
-                    "collaborations"
-                }, HTTPStatus.BAD_REQUEST
-            if collaboration_id in collabs_with_algo_store:
-                return {
-                    "msg": "Algorithm store is already available for this "
-                    "collaboration"
-                }, HTTPStatus.BAD_REQUEST
-            if not collaboration_id:
-                # algorithm store is currently available for some
-                # collaborations, but now it will be available for all of them.
-                # Remove the records that only make it available to some
-                # collaborations (this prevents duplicates)
-                records_to_delete = existing_algorithm_stores
-
-        # raise a warning if the algorithm store url is insecure (i.e.
-        # localhost)
-        force = data.get("force", False)
-        if not force and (
-            "localhost" in algorithm_store_url or "127.0.0.1" in algorithm_store_url
-        ):
-            return {
-                "msg": "Algorithm store url is insecure: localhost services "
-                "may be run on any computer. Add it anyway by setting the "
-                "'force' flag to true, but only do so for development servers!"
-            }, HTTPStatus.BAD_REQUEST
-
-        server_url = cls._get_server_url(config, data.get("server_url"))
-        if not server_url:
-            return {
-                "msg": "The 'server_url' key is required in the server "
-                "configuration, or as a parameter. Please add it or ask your "
-                "server administrator to specify it in the server configuration."
-            }, HTTPStatus.BAD_REQUEST
-
-        # whitelist this vantage6 server url for the algorithm store
-        response, status = cls._request_algo_store(
-            algo_store_url=algorithm_store_url,
-            server_url=server_url,
-            endpoint="vantage6-server",
-            method="post",
-            force=force,
-            headers=headers,
         )
         if status != HTTPStatus.CREATED:
             return response, status
-
-        # delete and create records
-        for record in records_to_delete:
-            record.delete()
-        algorithm_store = db.AlgorithmStore(
-            name=data["name"],
-            url=algorithm_store_url,
-            collaboration_id=collaboration_id,
-        )
-        algorithm_store.save()
-
-        return algorithm_store_schema.dump(algorithm_store), HTTPStatus.CREATED
+        else:
+            return algorithm_store_schema.dump(response), status
 
 
 class AlgorithmStore(AlgorithmStoreBase):
@@ -836,15 +565,13 @@ class AlgorithmStore(AlgorithmStoreBase):
         if len(other_algorithm_stores) == 1:
             # only this algorithm store uses this url, so delete the
             # whitelisting
-            server_url = self._get_server_url(
-                self.config, request.args.get("server_url")
-            )
+            server_url = get_server_url(self.config, request.args.get("server_url"))
             if not server_url:
                 return {
                     "msg": "The 'server_url' query parameter is required"
                 }, HTTPStatus.BAD_REQUEST
             # get the ID of the whitelisted server, then delete it
-            response, status = self._request_algo_store(
+            response, status = request_algo_store(
                 algorithm_store.url,
                 server_url,
                 endpoint="vantage6-server",
@@ -870,7 +597,7 @@ class AlgorithmStore(AlgorithmStoreBase):
             # remove all linked servers with the given url
             for server in result:
                 server_id = server["id"]
-                response, status = self._request_algo_store(
+                response, status = request_algo_store(
                     algorithm_store.url,
                     server_url,
                     endpoint=f"vantage6-server/{server_id}",
